+++ frontend/src/pages/installmentlist.js
@@ -1,25 +1,25 @@
 import AppLayout from '@/components/Layouts/AppLayout'
-import DashboardReport from '@/components/Report/DashboardReport'
+import InstallmentList from "@/components/installment/InstallmentList"
 import Head from 'next/head'
 
-const Dashboard = () => {
+
+
+const Borrower = () => {
     return (
         <AppLayout
             header={
                 <h2 className="font-semibold text-xl text-gray-800 leading-tight">
-                    Dashboard
+                    Borrower
                 </h2>
             }>
-
             <Head>
-                <title>Finace - Dashboard</title>
+                <title>Finace - Add Borrower</title>
             </Head>
-
             <div className="py-12">
                 <div className="max-w-7xl mx-auto sm:px-6 lg:px-8">
                     <div className="bg-white overflow-hidden shadow-sm sm:rounded-lg">
                         <div className="p-6 bg-white border-b border-gray-200">
-                           <DashboardReport></DashboardReport>
+                          <InstallmentList ></InstallmentList>
                         </div>
                     </div>
                 </div>
@@ -28,4 +28,4 @@ const Dashboard = () => {
     )
 }
 
-export default Dashboard
+export default Borrower;
